Update userId data type and fix FeignClient parameter
Changed the data type of userId from Integer to Long across multiple files for data consistency. Fixed the wrong usage of FeignClient annotation in ProductOrderFeign by changing the parameter from "@FeignClient("order-service")" to "@FeignClient(name="order-service")". Also, corrected a typo in the RabbitListener annotation in the ProductStockMQListener file.

+++ coupon-service/src/main/java/tech/qiuweihong/mapper/CouponRecordMapper.java
@@ -17,7 +17,7 @@ import java.util.List;
 public interface CouponRecordMapper extends BaseMapper<CouponRecordDO> {
 
 
-    int lockCouponState(@Param("userId")Integer id, @Param("useState")String useState, @Param("couponList")List<Long> couponList);
+    int lockCouponState(@Param("userId")Long id, @Param("useState")String useState, @Param("couponList")List<Long> couponList);
 
     void updateState(@Param("couponRecordId") Long couponRecordId,@Param("state") String name);
 }
